Authentication wire-ups

+++ packages/express-backend/src/server.js
@@ -3,6 +3,7 @@ import express from 'express'
 import bodyParser from 'body-parser'
 
 import Database from '../src/persistence/mongoose-connection.js'
+import Passwords from './handler/passwords.js'
 
 import { applyKeywordItemSearch } from './handler/search-filter.js'
 
@@ -26,6 +27,9 @@ app.get('/users', async (req, res) => {
 })
 
 app.post('/users', async (req, res) => {
+  const user = req.body
+  user['hashed_password'] = await Passwords.hashPassword(req.body.password)
+
   const createdUser = await Database.createUser(req.body)
 
   const statusCode = createdUser === undefined ? 400 : 201
@@ -44,16 +48,38 @@ app.get('/users/:id', async (req, res) => {
 
 app.delete('/users/:userId', async (req, res) => {
   const userId = req.params.userId
-  await Database.deleteUser(userId)
 
-  res.status(204).send()
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      await Database.deleteUser(userId)
+      res.status(204).send()
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.delete('/items/:itemId', async (req, res) => {
   const itemId = req.params.itemId
-  await Database.deleteItem(itemId)
+  const item = Database.findItemById(itemId)
 
-  res.status(204).send()
+  // If the item does not exist, return success
+  if (item === undefined) {
+    res.status(204).send()
+    return
+  }
+
+  const userId = item.seller_id
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      await Database.deleteItem(itemId)
+      res.status(204).send()
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.get('/items', async (req, res) => {
@@ -87,12 +113,20 @@ app.get('/users/:userId/wishlist', async (req, res) => {
 
 app.put('/users/:userId/wishlist/:itemId', async (req, res) => {
   const userId = req.params.userId
-  const itemId = req.params.itemId
 
-  const editedWishlist = await Database.addToWishList(userId, itemId)
-  const statusCode = editedWishlist === undefined ? 500 : 200
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      const itemId = req.params.itemId
 
-  res.status(statusCode).send(editedWishlist)
+      const editedWishlist = await Database.addToWishList(userId, itemId)
+      const statusCode = editedWishlist === undefined ? 500 : 200
+
+      res.status(statusCode).send(editedWishlist)
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.delete('/users/:userId/wishlist/:itemId', async (req, res) => {
@@ -104,12 +138,31 @@ app.delete('/users/:userId/wishlist/:itemId', async (req, res) => {
   res.status(204).send()
 })
 
+app.post('/login', async (req, res) => {
+  const { username } = req.body
+  const user = await Database.findUserByUsername(username)
+
+  if (!user) {
+    res.status(404).end()
+    return
+  }
+
+  Passwords.loginUser(req, res, user)
+})
+
 app.post('/users/:userId/items', async (req, res) => {
   const userId = req.params.userId
-  const createdItem = await Database.createItem(req.body, userId)
-  const statusCode = createdItem === undefined ? 500 : 201
 
-  res.status(statusCode).send(createdItem || {})
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      const createdItem = await Database.createItem(req.body, userId)
+      const statusCode = createdItem === undefined ? 500 : 201
+      res.status(statusCode).send(createdItem || {})
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.get('/users/:userId/items', async (req, res) => {
@@ -129,22 +182,41 @@ app.get('/users/:userId/items-bought', async (req, res) => {
 })
 
 app.patch('/users/:sellerId/items/:itemId', async (req, res) => {
-  const itemId = req.params.itemId
-  const buyerId = req.body.buyerId
+  const sellerId = req.params.sellerId
 
-  if (!buyerId) {
-    return res.status(400).send({ error: 'buyerId is required' })
-  }
+  Passwords.authenticateUser(req, sellerId)
+    .then(async () => {
+      const itemId = req.params.itemId
+      const buyerId = req.body.buyerId
 
-  try {
-    const updatedItem = await Database.updateItemBuyerAndAddToUserBought(
-      itemId,
-      buyerId
-    )
-    res.status(200).send(updatedItem)
-  } catch (error) {
-    res.status(500).send({ error: error.message })
-  }
+      if (!buyerId) {
+        return res.status(400).send({ error: 'buyerId is required' })
+      }
+
+      try {
+        const updatedItem = await Database.updateItemBuyerAndAddToUserBought(
+          itemId,
+          buyerId
+        )
+        res.status(200).send(updatedItem)
+      } catch (error) {
+        res.status(500).send({ error: error.message })
+      }
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
+})
+
+app.get('/users/usernames/:username', async (req, res) => {
+  const { username } = req.params
+
+  const user = await Database.findUserByUsername(username)
+
+  const statusCode = user === undefined ? 404 : 200
+
+  res.status(statusCode).send(user)
 })
 
 app.listen(process.env.PORT || port, () => {
